pyproject.toml
[project]
name = "miniagent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.3.0",
    "duckduckgo-search>=6.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.25.0",
]
